refactor: revert internal re-exporting index files
