+++ services/data/src/engine/helpers/getMutationFetchType.ts
@@ -1,4 +1,5 @@
-import { Mutation, FetchType } from '../types'
+import { FetchType } from '../types/ExecuteOptions'
+import { Mutation } from '../types/Mutation'
 
 export const getMutationFetchType = (mutation: Mutation): FetchType =>
     mutation.type === 'update'
