--- services/data/src/engine/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from './getMutationFetchType'
-export * from './resolveDynamicQuery'
