+++ services/data/src/engine/helpers/resolveDynamicQuery.test.ts
@@ -1,5 +1,5 @@
 import { resolveDynamicQuery } from './resolveDynamicQuery'
-import { QueryVariables } from '../types'
+import { QueryVariables } from '../types/Query'
 
 describe('resolveDynamicQuery', () => {
     it('Should return an unmodified query if no dynamic properties exist', () => {
