+++ services/data/src/engine/helpers/resolveDynamicQuery.ts
@@ -1,4 +1,8 @@
-import { QueryVariables, ResourceQuery, ResolvedResourceQuery } from '../types'
+import {
+    QueryVariables,
+    ResourceQuery,
+    ResolvedResourceQuery,
+} from '../types/Query'
 
 export const resolveDynamicQuery = (
     { resource, id, data, params }: ResourceQuery,
