+++ services/data/src/engine/index.ts
@@ -1,2 +1,9 @@
 export * from './DataEngine'
-export * from './types'
+export * from './types/DataEngineLink'
+export * from './types/ExecuteOptions'
+export * from './types/FetchError'
+export * from './types/JsonValue'
+export * from './types/Mutation'
+export * from './types/PossiblyDynamic'
+export * from './types/Query'
+export * from './types/QueryParameters'
