--- services/data/src/engine/types/index.ts
@@ -1,8 +1,0 @@
-export * from './DataEngineLink'
-export * from './ExecuteOptions'
-export * from './FetchError'
-export * from './JsonValue'
-export * from './Mutation'
-export * from './PossiblyDynamic'
-export * from './Query'
-export * from './QueryParameters'
