--- services/data/src/links/RestAPILink/index.ts
@@ -1,4 +1,0 @@
-export * from './fetchData'
-export * from './path'
-export * from './queryToRequestOptions'
-export * from './queryToResourcePath'
