--- services/data/src/react/components/index.ts
@@ -1,4 +1,0 @@
-export * from './CustomDataProvider'
-export * from './DataMutation'
-export * from './DataProvider'
-export * from './DataQuery'
