--- services/data/src/react/context/index.ts
@@ -1,2 +1,0 @@
-export * from './DataContext'
-export * from './defaultContext'
