+++ services/data/src/react/hooks/useDataEngine.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import { DataContext } from '../context'
+import { DataContext } from '../context/DataContext'
 
 export const useDataEngine = () => {
     const context = useContext(DataContext)
